feat: enhance model type refactor
