+++ camel/models/__init__.py
@@ -21,7 +21,6 @@ from .mistral_model import MistralModel
 from .model_factory import ModelFactory
 from .nemotron_model import NemotronModel
 from .ollama_model import OllamaModel
-from .open_source_model import OpenSourceModel
 from .openai_audio_models import OpenAIAudioModels
 from .openai_compatible_model import OpenAICompatibleModel
 from .openai_model import OpenAIModel
@@ -41,7 +40,6 @@ __all__ = [
     'GroqModel',
     'StubModel',
     'ZhipuAIModel',
-    'OpenSourceModel',
     'ModelFactory',
     'LiteLLMModel',
     'OpenAIAudioModels',
