+++ camel/models/model_factory.py
@@ -21,7 +21,6 @@ from camel.models.groq_model import GroqModel
 from camel.models.litellm_model import LiteLLMModel
 from camel.models.mistral_model import MistralModel
 from camel.models.ollama_model import OllamaModel
-from camel.models.open_source_model import OpenSourceModel
 from camel.models.openai_compatible_model import OpenAICompatibleModel
 from camel.models.openai_model import OpenAIModel
 from camel.models.reka_model import RekaModel
@@ -30,8 +29,7 @@ from camel.models.stub_model import StubModel
 from camel.models.togetherai_model import TogetherAIModel
 from camel.models.vllm_model import VLLMModel
 from camel.models.zhipuai_model import ZhipuAIModel
-from camel.types import ModelPlatformType, ModelType
-from camel.types.inner_model_type import InnerModelType
+from camel.types import InnerModelType, ModelPlatformType, ModelType
 from camel.utils import BaseTokenCounter
 
 
@@ -62,7 +60,7 @@ class ModelFactory:
                 into the backend constructor. (default: :obj:`None`)
             token_counter (Optional[BaseTokenCounter], optional): Token
                 counter to use for the model. If not provided,
-                :obj:`OpenAITokenCounter(PredefinedModelType.GPT_4O_MINI)`
+                :obj:`OpenAITokenCounter(ModelType.GPT_4O_MINI)`
                 will be used if the model platform didn't provide official
                 token counter. (default: :obj:`None`)
             api_key (Optional[str], optional): The API key for authenticating
@@ -79,30 +77,26 @@ class ModelFactory:
         model_class: Optional[Type[BaseModelBackend]] = None
         model = InnerModelType(model_type)
 
-        if model.is_open_source:
-            if model_platform.is_ollama:
-                return OllamaModel(
-                    model, model_config_dict, url, token_counter
-                )
-            elif model_platform.is_vllm:
-                model_class = VLLMModel
-            elif model_platform.is_openai_compatible_model:
-                model_class = OpenAICompatibleModel
-            elif model_platform.is_samba:
-                model_class = SambaModel
-            elif model_platform.is_together:
-                model_class = TogetherAIModel
-            elif model_platform.is_open_source:
-                model_class = OpenSourceModel
+        if model_platform.is_ollama:
+            return OllamaModel(model, model_config_dict, url, token_counter)
+        elif model_platform.is_vllm:
+            model_class = VLLMModel
+        elif model_platform.is_openai_compatible_model:
+            model_class = OpenAICompatibleModel
+        elif model_platform.is_samba:
+            model_class = SambaModel
+        elif model_platform.is_together:
+            model_class = TogetherAIModel
         elif model_platform.is_litellm:
             model_class = LiteLLMModel
+
         elif model_platform.is_openai and model.is_openai:
             model_class = OpenAIModel
         elif model_platform.is_azure and model.is_azure_openai:
             model_class = AzureOpenAIModel
         elif model_platform.is_anthropic and model.is_anthropic:
             model_class = AnthropicModel
-        elif model_platform.is_groq:
+        elif model_platform.is_groq and model.is_groq:
             model_class = GroqModel
         elif model_platform.is_zhipuai and model.is_zhipuai:
             model_class = ZhipuAIModel
