+++ camel/models/ollama_model.py
@@ -42,7 +42,7 @@ class OllamaModel(BaseModelBackend):
             (default: :obj:`None`)
         token_counter (Optional[BaseTokenCounter], optional): Token counter to
             use for the model. If not provided, :obj:`OpenAITokenCounter(
-            PredefinedModelType.GPT_4O_MINI)` will be used.
+            ModelType.GPT_4O_MINI)` will be used.
             (default: :obj:`None`)
 
     References:
