+++ camel/models/samba_model.py
@@ -63,7 +63,7 @@ class SambaModel(BaseModelBackend):
             sambanova.ai/v1`)
         token_counter (Optional[BaseTokenCounter], optional): Token counter to
             use for the model. If not provided, :obj:`OpenAITokenCounter(
-            PredefinedModelType.GPT_4O_MINI)` will be used.
+            ModelType.GPT_4O_MINI)` will be used.
     """
 
     def __init__(
