+++ camel/models/togetherai_model.py
@@ -50,7 +50,7 @@ class TogetherAIModel(BaseModelBackend):
             (default: :obj:`None`)
         token_counter (Optional[BaseTokenCounter], optional): Token counter to
             use for the model. If not provided, :obj:`OpenAITokenCounter(
-            PredefinedModelType.GPT_4O_MINI)` will be used.
+            ModelType.GPT_4O_MINI)` will be used.
     """
 
     def __init__(
