+++ camel/models/vllm_model.py
@@ -46,7 +46,7 @@ class VLLMModel(BaseModelBackend):
             (default: :obj:`None`)
         token_counter (Optional[BaseTokenCounter], optional): Token counter to
             use for the model. If not provided, :obj:`OpenAITokenCounter(
-            PredefinedModelType.GPT_4O_MINI)` will be used.
+            ModelType.GPT_4O_MINI)` will be used.
             (default: :obj:`None`)
 
     References:
