+++ camel/models/zhipuai_model.py
@@ -48,7 +48,7 @@ class ZhipuAIModel(BaseModelBackend):
             (default: :obj:`https://open.bigmodel.cn/api/paas/v4/`)
         token_counter (Optional[BaseTokenCounter], optional): Token counter to
             use for the model. If not provided, :obj:`OpenAITokenCounter(
-            PredefinedModelType.GPT_4O_MINI)` will be used.
+            ModelType.GPT_4O_MINI)` will be used.
             (default: :obj:`None`)
     """
