+++ camel/utils/__init__.py
@@ -43,7 +43,6 @@ from .token_counting import (
     LiteLLMTokenCounter,
     MistralTokenCounter,
     OpenAITokenCounter,
-    OpenSourceTokenCounter,
     get_model_encoding,
 )
 
@@ -60,7 +59,6 @@ __all__ = [
     'get_model_encoding',
     'BaseTokenCounter',
     'OpenAITokenCounter',
-    'OpenSourceTokenCounter',
     'LiteLLMTokenCounter',
     'Constants',
     'text_extract_from_web',
