+++ camel/utils/token_counting.py
@@ -45,150 +45,6 @@ SQUARE_TOKENS = 170
 EXTRA_TOKENS = 85
 
 
-def messages_to_prompt(
-    messages: List[OpenAIMessage], model: InnerModelType
-) -> str:
-    r"""Parse the message list into a single prompt following model-specific
-    formats.
-
-    Args:
-        messages (List[OpenAIMessage]): Message list with the chat history
-            in OpenAI API format.
-        model (InnerModelType): Model type for which messages will be parsed.
-
-    Returns:
-        str: A single prompt summarizing all the messages.
-    """
-    system_message = messages[0]["content"]
-
-    ret: str
-    if model in [
-        ModelType.LLAMA_2,
-        ModelType.LLAMA_3,
-        ModelType.GROQ_LLAMA_3_8B,
-        ModelType.GROQ_LLAMA_3_70B,
-    ]:
-        # reference: https://github.com/facebookresearch/llama/blob/cfc3fc8c1968d390eb830e65c63865e980873a06/llama/generation.py#L212
-        seps = [" ", " </s><s>"]
-        role_map = {"user": "[INST]", "assistant": "[/INST]"}
-
-        system_prompt = f"[INST] <<SYS>>\n{system_message}\n<</SYS>>\n\n"
-        ret = ""
-        for i, msg in enumerate(messages[1:]):
-            role = role_map[msg["role"]]
-            content = msg["content"]
-            if content:
-                if not isinstance(content, str):
-                    raise ValueError(
-                        "Currently multimodal context is not "
-                        "supported by the token counter."
-                    )
-                if i == 0:
-                    ret += system_prompt + content
-                else:
-                    ret += role + " " + content + seps[i % 2]
-            else:
-                ret += role
-        return ret
-    elif model in [ModelType.VICUNA, ModelType.VICUNA_16K]:
-        seps = [" ", "</s>"]
-        role_map = {"user": "USER", "assistant": "ASSISTANT"}
-
-        system_prompt = f"{system_message}"
-        ret = system_prompt + seps[0]
-        for i, msg in enumerate(messages[1:]):
-            role = role_map[msg["role"]]
-            content = msg["content"]
-            if not isinstance(content, str):
-                raise ValueError(
-                    "Currently multimodal context is not "
-                    "supported by the token counter."
-                )
-            if content:
-                ret += role + ": " + content + seps[i % 2]
-            else:
-                ret += role + ":"
-        return ret
-    elif model == ModelType.GLM_4_OPEN_SOURCE:
-        system_prompt = f"[gMASK]<sop><|system|>\n{system_message}"
-        ret = system_prompt
-        for msg in messages[1:]:
-            role = msg["role"]
-            content = msg["content"]
-            if not isinstance(content, str):
-                raise ValueError(
-                    "Currently multimodal context is not "
-                    "supported by the token counter."
-                )
-            if content:
-                ret += "<|" + role + "|>" + "\n" + content
-            else:
-                ret += "<|" + role + "|>" + "\n"
-        return ret
-    elif model == ModelType.QWEN_2:
-        system_prompt = f"<|im_start|>system\n{system_message}<|im_end|>"
-        ret = system_prompt + "\n"
-        for msg in messages[1:]:
-            role = msg["role"]
-            content = msg["content"]
-            if not isinstance(content, str):
-                raise ValueError(
-                    "Currently multimodal context is not "
-                    "supported by the token counter."
-                )
-            if content:
-                ret += (
-                    '<|im_start|>'
-                    + role
-                    + '\n'
-                    + content
-                    + '<|im_end|>'
-                    + '\n'
-                )
-            else:
-                ret += '<|im_start|>' + role + '\n'
-        return ret
-    elif model == ModelType.GROQ_MIXTRAL_8_7B:
-        # Mistral/Mixtral format
-        system_prompt = f"<s>[INST] {system_message} [/INST]\n"
-        ret = system_prompt
-
-        for msg in messages[1:]:
-            if msg["role"] == "user":
-                ret += f"[INST] {msg['content']} [/INST]\n"
-            elif msg["role"] == "assistant":
-                ret += f"{msg['content']}</s>\n"
-
-            if not isinstance(msg['content'], str):
-                raise ValueError(
-                    "Currently multimodal context is not "
-                    "supported by the token counter."
-                )
-
-        return ret.strip()
-    elif model in [
-        ModelType.GROQ_GEMMA_7B_IT,
-        ModelType.GROQ_GEMMA_2_9B_IT,
-    ]:
-        # Gemma format
-        ret = f"<bos>{system_message}\n"
-        for msg in messages:
-            if msg["role"] == "user":
-                ret += f"Human: {msg['content']}\n"
-            elif msg["role"] == "assistant":
-                ret += f"Assistant: {msg['content']}\n"
-
-            if not isinstance(msg['content'], str):
-                raise ValueError(
-                    "Currently multimodal context is not supported by the token counter."
-                )
-
-        ret += "<eos>"
-        return ret
-    else:
-        raise ValueError(f"Invalid model type: {model}")
-
-
 def get_model_encoding(value_for_tiktoken: str):
     r"""Get model encoding from tiktoken.
 
@@ -229,58 +85,6 @@ class BaseTokenCounter(ABC):
             int: Number of tokens in the messages.
         """
         pass
-
-
-class OpenSourceTokenCounter(BaseTokenCounter):
-    def __init__(self, model_type: InnerModelType, model_path: str):
-        r"""Constructor for the token counter for open-source models.
-
-        Args:
-            model_type (InnerModelType): Model type for which tokens will be
-                counted.
-            model_path (str): The path to the model files, where the tokenizer
-                model should be located.
-        """
-
-        # Use a fast Rust-based tokenizer if it is supported for a given model.
-        # If a fast tokenizer is not available for a given model,
-        # a normal Python-based tokenizer is returned instead.
-        from transformers import AutoTokenizer
-
-        try:
-            tokenizer = AutoTokenizer.from_pretrained(
-                model_path,
-                use_fast=True,
-            )
-        except TypeError:
-            tokenizer = AutoTokenizer.from_pretrained(
-                model_path,
-                use_fast=False,
-            )
-        except Exception:
-            raise ValueError(
-                f"Invalid `model_path` ({model_path}) is provided. "
-                "Tokenizer loading failed."
-            )
-
-        self.tokenizer = tokenizer
-        self.model_type = model_type
-
-    def count_tokens_from_messages(self, messages: List[OpenAIMessage]) -> int:
-        r"""Count number of tokens in the provided message list using
-        loaded tokenizer specific for this type of model.
-
-        Args:
-            messages (List[OpenAIMessage]): Message list with the chat history
-                in OpenAI API format.
-
-        Returns:
-            int: Number of tokens in the messages.
-        """
-        prompt = messages_to_prompt(messages, self.model_type)
-        input_ids = self.tokenizer(prompt).input_ids
-
-        return len(input_ids)
 
 
 class OpenAITokenCounter(BaseTokenCounter):
