+++ examples/test/test_ai_society_example.py
@@ -15,7 +15,6 @@ from mock import patch
 
 import examples.ai_society.role_playing
 import examples.function_call.role_playing_with_functions
-import examples.models.role_playing_with_open_source_model
 from camel.models import ModelFactory
 from camel.types import ModelPlatformType, ModelType
 
@@ -35,12 +34,5 @@ def test_ai_society_role_playing_example():
 def test_role_playing_with_function_example():
     with patch('time.sleep', return_value=None):
         examples.function_call.role_playing_with_functions.main(
-            chat_turn_limit=2
-        )
-
-
-def test_role_playing_with_open_source_model():
-    with patch('time.sleep', return_value=None):
-        examples.models.role_playing_with_open_source_model.main(
             chat_turn_limit=2
         )
