+++ test/models/test_azure_openai_model.py
@@ -25,7 +25,7 @@ import re
 
 import pytest
 
-from camel.configs import ChatGPTConfig, OpenSourceConfig
+from camel.configs import ChatGPTConfig
 from camel.models import AzureOpenAIModel, ModelFactory
 from camel.types import ModelPlatformType, ModelType
 from camel.utils import OpenAITokenCounter
@@ -73,11 +73,7 @@ def test_openai_model_create(model_type: ModelType):
 @pytest.mark.model_backend
 def test_openai_model_unexpected_argument():
     model_type = ModelType.GPT_4
-    model_config = OpenSourceConfig(
-        model_path="vicuna-7b-v1.5",
-        server_url="http://localhost:8000/v1",
-    )
-    model_config_dict = model_config.as_dict()
+    model_config_dict = {"model_path": "vicuna-7b-v1.5"}
 
     with pytest.raises(
         ValueError,
