+++ test/models/test_model_factory.py
@@ -18,7 +18,6 @@ from camel.configs import (
     ChatGPTConfig,
     GeminiConfig,
     OllamaConfig,
-    OpenSourceConfig,
 )
 from camel.models import ModelFactory
 from camel.models.stub_model import StubTokenCounter
@@ -26,7 +25,6 @@ from camel.types import ModelPlatformType, ModelType
 from camel.utils import (
     AnthropicTokenCounter,
     OpenAITokenCounter,
-    OpenSourceTokenCounter,
 )
 
 parametrize = pytest.mark.parametrize(
@@ -34,7 +32,6 @@ parametrize = pytest.mark.parametrize(
     [
         (ModelPlatformType.OPENAI, ModelType.GPT_3_5_TURBO),
         (ModelPlatformType.OPENAI, ModelType.GPT_4_TURBO),
-        (ModelPlatformType.OPEN_SOURCE, ModelType.STUB),
     ],
 )
 
@@ -62,7 +59,7 @@ parameterize_token_counter = pytest.mark.parametrize(
         # Test Stub model
         # Stub model uses StubTokenCounter as default
         (
-            ModelPlatformType.OPEN_SOURCE,
+            ModelPlatformType.OPENAI_COMPATIBLE_MODEL,
             ModelType.STUB,
             ChatGPTConfig().as_dict(),
             None,
@@ -70,7 +67,7 @@ parameterize_token_counter = pytest.mark.parametrize(
             None,
         ),
         (
-            ModelPlatformType.OPEN_SOURCE,
+            ModelPlatformType.OPENAI_COMPATIBLE_MODEL,
             ModelType.STUB,
             ChatGPTConfig().as_dict(),
             OpenAITokenCounter(ModelType.GPT_4),
@@ -95,30 +92,7 @@ parameterize_token_counter = pytest.mark.parametrize(
             OpenAITokenCounter,
             ModelType.GPT_3_5_TURBO,
         ),
-        # Test OpenSource model (take VICUNA as an example)
-        (
-            ModelPlatformType.OPEN_SOURCE,
-            ModelType.VICUNA,
-            OpenSourceConfig(
-                model_path="lmsys/vicuna-7b-v1.5",
-                server_url="http://localhost:8000/v1",
-            ).as_dict(),
-            None,
-            OpenSourceTokenCounter,
-            ModelType.VICUNA,
-        ),
-        (
-            ModelPlatformType.OPEN_SOURCE,
-            ModelType.VICUNA,
-            OpenSourceConfig(
-                model_path="lmsys/vicuna-7b-v1.5",
-                server_url="http://localhost:8000/v1",
-            ).as_dict(),
-            OpenAITokenCounter(ModelType.GPT_4),
-            OpenAITokenCounter,
-            ModelType.GPT_4,
-        ),
-        # Test OpenSource model (take VICUNA as an example)
+        # Test GEMINI model
         (
             ModelPlatformType.GEMINI,
             ModelType.GEMINI_1_5_FLASH,
